+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+<a name="3.0.8"></a>
+## [3.0.8](https://github.com/duyluonglc/lucid-mongo/compare/v3.0.7...v3.0.8) (2018-05-13)
+
+
+### Bug Fixes
+
+* **connection:** fix connectionString with cluster ([2e54be0](https://github.com/duyluonglc/lucid-mongo/commit/2e54be0))
+
+
+
+<a name="3.0.7"></a>
+## [3.0.7](https://github.com/duyluonglc/lucid-mongo/compare/v3.0.6...v3.0.7) (2018-05-02)
+
+
+
+<a name="3.0.6"></a>
+## [3.0.6](https://github.com/duyluonglc/lucid-mongo/compare/v3.0.5...v3.0.6) (2018-04-24)
+
+
+### Bug Fixes
+
+* **referMany:** fix can not detach ([f8ad672](https://github.com/duyluonglc/lucid-mongo/commit/f8ad672))
+
+
+
+<a name="3.0.5"></a>
+## [3.0.5](https://github.com/duyluonglc/lucid-mongo/compare/v3.0.4...v3.0.5) (2018-04-06)
+
+
+### Bug Fixes
+
+* **queryBuilder:** allow format field on query $and $or $nor ([c388dc0](https://github.com/duyluonglc/lucid-mongo/commit/c388dc0))
+
+
+
 <a name="3.0.4"></a>
 ## [3.0.4](https://github.com/duyluonglc/lucid-mongo/compare/v3.0.3...v3.0.4) (2018-03-30)